mysite/requirements.txt
astroid==2.3.3
autopep8==1.5
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.9
django-storages==1.8
djangorestframework==3.11.0
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.2.0
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.8
imutils==0.5.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==16.7.5
virtualenvwrapper-win==1.2.5
wrapt==1.11.2
